refactor: 태그 도메인 패키지 리팩토링
- 불필요한 static 필드 제거
- 메서드 순서 변경 및 사용하지 않는 메서드 제거
- 패키지를 통한 Jdbc, Jpa 레포지토리 분리
