+++ backend/src/main/java/com/example/backend/business/core/tag/entity/values/StudyHashTags.java
@@ -12,8 +12,6 @@ import java.util.Set;
 @Embeddable
 public class StudyHashTags {
 
-    private static final int MAX_HASH_TAG_COUNT = 8;
-
     @OneToMany(mappedBy = "study", cascade = {CascadeType.PERSIST, CascadeType.REMOVE}, orphanRemoval = true)
     private Set<StudyHashTag> studyHashTags;
 
@@ -29,6 +27,15 @@ public class StudyHashTags {
         this.studyHashTags = studyHashTags;
     }
 
+    private static void validateCreateHashTags(Set<StudyHashTag> hashTags) {
+        if (Objects.isNull(hashTags) || hashTags.isEmpty()) {
+            throw new IllegalArgumentException("해시태그 값을 입력해주세요.");
+        }
+        if (hashTags.size() > 8) {
+            throw new IllegalArgumentException("해시태그는 최대 8개까지 입력할 수 있습니다.");
+        }
+    }
+
     public static StudyHashTags initHashTags() {
         return new StudyHashTags();
     }
@@ -36,15 +43,6 @@ public class StudyHashTags {
     public static StudyHashTags from(Set<StudyHashTag> studyHashTags) {
         validateCreateHashTags(studyHashTags);
         return new StudyHashTags(studyHashTags);
-    }
-
-    private static void validateCreateHashTags(Set<StudyHashTag> hashTags) {
-        if (Objects.isNull(hashTags) || hashTags.isEmpty()) {
-            throw new IllegalArgumentException("해시태그 값을 입력해주세요.");
-        }
-        if (hashTags.size() > MAX_HASH_TAG_COUNT) {
-            throw new IllegalArgumentException("해시태그는 최대 8개까지 입력할 수 있습니다.");
-        }
     }
 
     public void add(StudyHashTag studyHashTag) {
@@ -59,7 +57,7 @@ public class StudyHashTags {
         if (hashTags.isEmpty()) {
             return;
         }
-        if (hashTags.size() > MAX_HASH_TAG_COUNT) {
+        if (hashTags.size() > 8) {
             throw new IllegalArgumentException("해시태그는 최대 8개까지 입력할 수 있습니다.");
         }
 
@@ -81,7 +79,7 @@ public class StudyHashTags {
     }
 
     private boolean isEmpty() {
-        return this.studyHashTags == null || this.studyHashTags.size() == 0;
+        return this.studyHashTags == null || this.studyHashTags.isEmpty();
     }
 
     public int size() {
@@ -89,7 +87,7 @@ public class StudyHashTags {
     }
 
     public void validate() {
-        if (this.studyHashTags == null || this.studyHashTags.size() == 0) {
+        if (this.studyHashTags == null || this.studyHashTags.isEmpty()) {
             throw new IllegalArgumentException("해시태그가 존재하지 않습니다.");
         }
     }
